adding and removing carousel items from dashboard now works again
git-svn-id: https://www.europeanalabs.eu/svn/europeana/trunk@245 3c7f12e8-d31e-4037-9ce9-7a8ea9e2c7c8

+++ database/src/main/java/eu/europeana/database/dao/DashboardDaoImpl.java
@@ -882,13 +882,15 @@ public class DashboardDaoImpl implements DashboardDao {
 	}
 
     @Transactional
-    public Boolean removeCarouselItem(Long savedItemId) {
-        // remove carousel item and give back a user
-        SavedItem savedItem = entityManager.getReference(SavedItem.class, savedItemId);
-        if (savedItem == null) {
-            throw new IllegalArgumentException("Unable to find saved item: " + savedItemId);
+    public Boolean removeCarouselItem(Long carouselItemId) {
+        CarouselItem carouselItem = entityManager.getReference(CarouselItem.class, carouselItemId);
+        if (carouselItem == null) {
+            throw new IllegalArgumentException("Unable to find saved item: " + carouselItemId);
         }
-        CarouselItem carouselItem = savedItem.getCarouselItem();
+        SavedItem savedItem = entityManager.getReference(SavedItem.class, carouselItem.getSavedItem().getId());
+        if (savedItem == null) {
+            throw new IllegalArgumentException("Unable to find saved item: " + carouselItemId);
+        }
         savedItem.setCarouselItem(null);
         entityManager.remove(carouselItem);
         entityManager.flush();
